chore: Update genmon version to 0.0.6 and Docker image to bigbeartechworld/big-bear-genmon:0.0.6

+++ genmon/entrypoint.sh
@@ -12,14 +12,25 @@ if [ ! -f /git/genmon/startgenmon.sh ]; then
   rm -rf /git/genmon/.git
 fi
 
+# Wait for the /git/genmon/startgenmon.sh to exist
+while [ ! -f /git/genmon/startgenmon.sh ]; do
+  echo "Waiting for /git/genmon/startgenmon.sh to be available..."
+  sleep 1
+done
+
 # Ensure the /etc/genmon directory exists and has the necessary configuration files
 if [ ! -f /etc/genmon/genmon.conf ]; then
   echo "Creating default configuration in /etc/genmon..."
   mkdir -p /etc/genmon
-  cat <<EOL > /etc/genmon/genmon.conf
+  if [ -f /git/genmon/conf/genmon.conf ]; then
+    cp /git/genmon/conf/genmon.conf /etc/genmon/genmon.conf
+  else
+    echo "Warning: /git/genmon/conf/genmon.conf does not exist."
+    cat <<EOL > /etc/genmon/genmon.conf
 # Default genmon configuration
 use_serial_tcp = $USE_SERIAL_TCP
 EOL
+  fi
 fi
 
 # Change the environment variable to use serial TCP in the genmon.conf file
